requirements.txt
streamlit==1.55.0
supabase==2.10.0
anthropic==0.86.0
pandas==2.3.3
openpyxl==3.1.5
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.32.5
watchdog==6.0.0
Pillow==11.2.1
